Initial Configuration for Book Shelf application
For the first stage

This will be a simple application to upload books to a web platform in there the
registered users are able to rate such books and also download them. Only the admin
user would be able to add new books the other users can only download them and rate
them.

In a second Stage:
We weill implement background jobs to retrieve books from Google's API using sidekiq
and redis.