Include new table and resources for Books
This includes the new model to handle all the Books attributes in a new table call
Books, includes the controller, the default generated assets, views, and controller
actions

This also includes all the CRUD actions and views for the books, so now the user is
to create, read, update and delete books.

Also include all the links to allow navigation between views.

This implements books in projects.