Include simple form and boostrap for faster front end development
We want to use the gem simple form to be able to handle form in an easier way than usual
and we want to use it along with boostrap for fast styling.

Simple form installation specifically for bootstrap; this includes all the necessary gems
for bootrstap and simple form to work together and correctly.