Add authentication to application
We want to be able to manage users in our application, meaning that we want a new user
to be able to create an acoount sign in to it and be able to rate books only. For this
we will use the gem Devise.

This includes all the initial configuration to include Devise in project including li-
braries, views, installation, and User model.

This implementes Devise in Book review project.