Add relation between users and books
We want to have a way to track which user created each book, and this way be
able to show or hide the edit and destroy links to actions in the views
by validating if the logged user was who created the current book.

This implements the database relationship between the user model and the
book model. It also implemnts conditions in the views to show or hide the
edit and destroy links.

Note: right now even if the user who is loged in wasnt who created the book
if the user types the right url to the edit action and correct id the system
will allow that user to make changes or even destroy the book in questing,
so in the next sprint please make sure to include some validations at the
model level if possible to validate this matter. Recommend to user gems such
as Pundit to delimite the scopes.