Implement categories for books
We want to be able to sort the books by categories that way the user will be
able to filter books by the desire category.

Generates the model, migrations, and relations between models, in this case
will be that a book belongs to a category and a category can have multiple
books.

This also includes the changes in the necessary views such as the header
partial, the index views, and in the forms so that now whenever a user
creates a new book it will be associated with a category.

This implements categories for books in the application.