Implement images for books to represent the cover
We want to be able to attrach a file for now to each book that we create
that way we can use this file as the cover image of the book.

For this we will use Active Storage which actually comes as default storage
for file attachement. For now this is the configuration only for local
development environment.

This implement active storage for image attachment in books only for
development environment.