Configure AWS S3 for image storage in production (Heroku)
We want to be able to guarantee that the images will remaing with the same
quality at all time and of course always available for when we need them.
For that matter we will use AWS S3 bucket services to store all uploaded
images by the users.

This implement the configuration, libraries, and setup for S3 bucket.