Implement review for books in web app
We want to allow the user to post a review and a rating to a book that they choose
the only requirement is that they are login.

This includes the relations between the models User, Book, and Review, including the
necessary views and partial.

When the user types a rating for a book, when the review is posted it displays the
comment and the number of start the user gave to that bookk.

Only the user who created the review is able to edit or remove that review.

This implements reviews and ratings.