Implement a more visual rating system
We want to be able to rate a book with the tipical 5 star animation for rating, for
that we will use a js library named Raty.js.

We needed to copy the javascript file under lib folder (jquery.raty.js) and also the
images for this to work (star-on, star-off and star-half)

Then we called the rigut classes in the right places and remove the rating input
because now everything is handle thru js.

Implements 5 star images rating system.