Implement new field for the book file
This is the file that will store the pdf, ebook, or whatever extension of the book
for us.

This also adds a set of new links to be able to preview the document on the website
or to download the content.

This implements the new field for the book file.