Include Sidekiq for asynchronous jobs
We want to use Redis for cache and Sidekiq for queue jobs.

This includes all the necessary gems for Redis and for Sidekiq, it also implements
all the necessary configuration on the config/aplication.rb file as well as all the
routes on the routes file.

This implements Sidekiq and Redis in project.