Include Rubocop in project
We want to use Rubocop in our project to guarantee that we will be able to use only
the good conventions and programming good practices in the development process.

This includes Rubocop in the project and also the configuration file to be able to
change the behave of certain cops and the files to be inspected.