Include seeds to load books example
We want to be able to export a bunch of books from a seed file this way we dont
have to create the books every time, instead we can simply run the seeds command
and have all the records available right the way.

This implements seeds and image files for the book's covers.