Implement pagination on books
We want to paginate the books index to have the website more organize.

For this we implemented a new gem we came across doing some research, the gem's
name is pagy. Pagy claims to be 19x times faster and 26x lighter that Kaminari
this is why we decided to implement it.

We added the gem to Gemfile include the Backend library in the application
controller and the Frontend library in the application helper.

This implements pagination on the project using the new gem Pagy.